extracting GreeterService to expose bug in finding beans with annotation
