+++ grpc-spring-boot-starter-demo/src/main/java/org/lognet/springboot/grpc/demo/DemoApp.java
@@ -6,6 +6,7 @@ import io.grpc.examples.CalculatorOuterClass;
 import org.lognet.springboot.grpc.GRpcService;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
 
 import io.grpc.examples.GreeterGrpc;
 import io.grpc.examples.GreeterOuterClass;
@@ -19,15 +20,10 @@ import io.grpc.stub.StreamObserver;
 @SpringBootApplication
 public class DemoApp {
 
-    @GRpcService(interceptors = { LogInterceptor.class })
-    public static class GreeterService extends GreeterGrpc.GreeterImplBase {
-        @Override
-        public void sayHello(GreeterOuterClass.HelloRequest request, StreamObserver<GreeterOuterClass.HelloReply> responseObserver) {
-            final GreeterOuterClass.HelloReply.Builder replyBuilder = GreeterOuterClass.HelloReply.newBuilder().setMessage("Hello " + request.getName());
-            responseObserver.onNext(replyBuilder.build());
-            responseObserver.onCompleted();
-        }
-    }
+    @Bean
+    public GreeterService greeterService() {
+		return new GreeterService();
+	}
 
     @GRpcService
     public static class CalculatorService extends CalculatorGrpc.CalculatorImplBase{
