+++ grpc-spring-boot-starter-demo/src/main/java/org/lognet/springboot/grpc/demo/GreeterService.java
@@ -1,0 +1,21 @@
+package org.lognet.springboot.grpc.demo;
+
+import io.grpc.examples.CalculatorGrpc;
+import io.grpc.examples.CalculatorOuterClass;
+import org.lognet.springboot.grpc.GRpcService;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+import io.grpc.examples.GreeterGrpc;
+import io.grpc.examples.GreeterOuterClass;
+import io.grpc.stub.StreamObserver;
+
+@GRpcService(interceptors = { LogInterceptor.class })
+public class GreeterService extends GreeterGrpc.GreeterImplBase {
+    @Override
+    public void sayHello(GreeterOuterClass.HelloRequest request, StreamObserver<GreeterOuterClass.HelloReply> responseObserver) {
+        final GreeterOuterClass.HelloReply.Builder replyBuilder = GreeterOuterClass.HelloReply.newBuilder().setMessage("Hello " + request.getName());
+        responseObserver.onNext(replyBuilder.build());
+        responseObserver.onCompleted();
+    }
+}
